requirements.txt
pyyaml
eventlet
paramiko
pudb